Moves reqs to a folder, moves profile to an arg.
