+++ MANIFEST.in
@@ -3,3 +3,4 @@ include *.txt
 include *.md
 include *.py
 recursive-include will/templates *
+recursive-include requirements/*
