+++ docs/backends/encryption.md
@@ -1,0 +1,1 @@
+# Encrypt all the things
