+++ docs/backends/generation.md
@@ -1,0 +1,1 @@
+# It's good to have options
