+++ docs/backends/pubsub.md
@@ -1,0 +1,1 @@
+# Go with the flow
