+++ requirements.txt
@@ -1,7 +1,7 @@
 hiredis==0.2.0
 redis==2.10.6
 
--r requirements.base.txt
--r requirements.slack.txt
--r requirements.hipchat.txt
--r requirements.rocketchat.txt
+-r requirements/base.txt
+-r requirements/slack.txt
+-r requirements/hipchat.txt
+-r requirements/rocketchat.txt
