+++ requirements/couchbase.txt
@@ -1,2 +1,2 @@
 -e git+git://github.com/couchbase/couchbase-python-client#egg=couchbase-python-client
--r requirements.base.txt
+-r base.txt
