+++ requirements/hipchat.txt
