+++ requirements/zeromq.txt
@@ -1,0 +1,2 @@
+pyzmq==16.0.2
+-r base.txt
